// src/lib/types.ts
export enum GenderType {
  Female = 2,
  Male = 1,
  Other = 0,
}
// ...
export enum MonthName {
  NotSet = 0,
  January = 1,
  February = 2,
  March = 3,
  April = 4,
  May = 5,
  June = 6,
  July = 7,
  August = 8,
  September = 9,
  October = 10,
  November = 11,
  December = 12,
}